% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxpreps_indv_leaders.R
\name{maxpreps_indv_leaders}
\alias{maxpreps_indv_leaders}
\title{Get individual leaders for various sports and categories}
\usage{
maxpreps_indv_leaders(sport, category, state, class_year, position)
}
\arguments{
\item{sport}{string}

\item{category}{string, will vary by sport}

\item{state}{string, two character state abbreviation, e.g. 'CA'}

\item{class_year}{alphanumeric selection of 9-12. 9 = Freshmen, 10 = Sophomore, etc.}

\item{position}{string, will vary by sport}
}
\value{
data frame
}
\description{
Scrape individual leaderboards from www.maxpreps.com
}
\examples{
maxpreps_indv_leaders("basketball", "scoring", "ca", "12", "pg")
}

