% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxpreps_team_leaders.R
\name{maxpreps_team_leaders}
\alias{maxpreps_team_leaders}
\title{Get team leaders for various sports and categories}
\usage{
maxpreps_team_leaders(sport, category, state)
}
\arguments{
\item{sport}{string}

\item{category}{string, will vary by sport}

\item{state}{string, two character state abbreviation, e.g. 'CA'}
}
\value{
data frame
}
\description{
Scrape team leaderboards from www.maxpreps.com
}
\examples{
maxpreps_team_leaders("basketball", "scoring", "ca")
}

