% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxpreps_team_stats.R
\name{maxpreps_team_stats}
\alias{maxpreps_team_stats}
\title{Get team roster and statistics if available}
\usage{
maxpreps_team_stats(school, mascot, city, state, sport, year)
}
\arguments{
\item{school}{string}

\item{mascot}{string}

\item{city}{string}

\item{state}{string, two character state abbreviation, e.g. 'CA'}

\item{sport}{string}

\item{year}{string, in the format: 06-07, 09-10, etc.}
}
\value{
data frame
}
\description{
Scrape the specified team's roster and statistics pages from www.maxpreps.com
}
\examples{
maxpreps_team_stats("pacific union college prep", "falcons", "angwin", "ca", "basketball", "04-05")
}

